+++ include/pg_tre/pg_tre.h
@@ -49,7 +49,7 @@
 #define PG_TRE_FORMAT_VERSION PG_TRE_FORMAT_VERSION_LATEST
 
 /* String version returned by tre_version(). */
-#define PG_TRE_VERSION_STRING "pg_tre 1.5.4"
+#define PG_TRE_VERSION_STRING "pg_tre 1.5.5"
 
 /* Module GUCs, defined in src/module.c. */
 extern int  pg_tre_default_max_cost;